gpu_model_runner: match target-ViT encoder dtype to engine dtype
The TargetVisionEncoder construction at line 5471 hardcoded
`dtype=torch.bfloat16`. When the surrounding engine ran at float16
(SYS4 production regime), this produced bf16-dtype pre-encoded
`image_embeds` that the target's fp16 LLM cast back to fp16 on
ingest — small dtype drift that, combined with the HF-vs-vLLM
encoder-class mismatch we just fixed in paper_explore's
`target_vision.py`, caused 36% of cascade REGENs to return
empty/whitespace text or fall into degenerate token loops (📐 emoji
repetition on chartqa/mathvista records; immediate EOS on others).
SYS4 Phase A smoke A/B test (cascade vs raw-image regen on the same
10 records) confirmed 10/10 identical responses after both fixes —
previously the cascade response style/length differed from raw on
all 10 records even when raw was correct.

Use `self.dtype` (=`self.model_config.dtype`) so the encoder always
matches the engine. For multi-engine setups (draft fp16 + target
bf16, etc.) the encoder dtype tracks the *draft* engine, on the
assumption the user keeps both engines consistent.

+++ vllm/v1/worker/gpu_model_runner.py
@@ -5468,7 +5468,13 @@ class GPUModelRunner(
         self._target_vision = TargetVisionEncoder(
             target_model_id=self._target_vision_model_id,
             device=self.device,
-            dtype=torch.bfloat16,
+            # Match the engine's dtype so pre-encoded embeds drop into
+            # the target's LLM at its native precision. Previously
+            # hardcoded to bfloat16, which mismatched the target's
+            # float16 dtype and produced numerically drifted embeds
+            # — surfaced as degenerate REGEN responses
+            # (📐-emoji loops, immediate EOS) in SYS4 Phase B.
+            dtype=self.dtype,
             # Step B: dedicated stream so the encode's TP collectives
             # can overlap with the next main-forward step on the default
             # stream. NCCL ordering: all 4 ranks call collectives in the
